+++ lib/src/elements/sink/muxer/mod.rs
@@ -1,6 +1,7 @@
 mod file_muxer;
 mod hls_muxer;
 mod rtmp_muxer;
+mod rtsp_muxer;
 mod segmented_file_muxer;
 
 pub use file_muxer::{FileMuxer, FileMuxerError, FileMuxerStreamSink};
@@ -8,4 +9,5 @@ pub use hls_muxer::{
     HlsMode, HlsMuxer, HlsMuxerError, HlsMuxerStreamSink, HlsOptions, HlsSegmentFormat,
 };
 pub use rtmp_muxer::{RtmpMuxer, RtmpMuxerError, RtmpMuxerStreamSink};
+pub use rtsp_muxer::{RtspMuxer, RtspMuxerError, RtspMuxerStreamSink};
 pub use segmented_file_muxer::{SegmentPolicy, SegmentedFileMuxer};